+++ lib/webserver/apihandlers.py
@@ -1,4 +1,3 @@
-import urllib.parse
 from json import JSONDecodeError
 from urllib.parse import unquote
 
@@ -11,8 +10,6 @@ from lib.util import safe_json_dumps
 from lib.util.util import implement_fields_query
 from lib.webserver.auth import EnodoAuth
 from lib.webserver.basehandler import BaseHandler
-
-from enodo.jobs import JOB_TYPE_IDS
 
 auth = BasicAuthMiddleware(username=None, password=None, force=False)
 
@@ -35,7 +32,12 @@ class ApiHandlers:
             dict with siridb response
         """
         series_name = unquote(request.match_info['series_name'])
-        pool_id = int(request.rel_url.query.get('poolID', 0))
+        try:
+            pool_id = int(request.rel_url.query.get('poolID'))
+        except Exception:
+            return web.json_response(
+                {'error': "invalid pool_id"},
+                dumps=safe_json_dumps, status=400)
         response_output_id = request.rel_url.query.get('responseOutputID')
         if response_output_id is None:
             resp, status = {'error': 'Invalid responseOutputID'}, 400
@@ -60,18 +62,46 @@ class ApiHandlers:
             dict with state response
         """
         series_name = unquote(request.match_info['series_name'])
-        pool_id = int(request.rel_url.query.get('poolID', 0))
 
-        job_type = unquote(request.match_info['job_type'])
-        if job_type not in JOB_TYPE_IDS:
+        try:
+            pool_id = int(request.match_info['pool_id'])
+        except Exception:
             return web.json_response(
-                {"data": False}, dumps=safe_json_dumps, status=400)
+                {"error": "Invalid pool_id"},
+                dumps=safe_json_dumps, status=400)
+
+        try:
+            job_type_id = int(request.match_info['job_type_id'])
+        except Exception:
+            return web.json_response(
+                {"error": "Invalid job_type_id"},
+                dumps=safe_json_dumps, status=400)
+
         data, status = await BaseHandler.resp_query_series_state(
             pool_id,
             series_name,
-            JOB_TYPE_IDS[job_type])
+            job_type_id)
         return web.json_response(
             data, dumps=safe_json_dumps, status=status)
+
+    @classmethod
+    @EnodoAuth.auth.required
+    async def query_worker_stats(cls, request):
+        """Query stats of workers in a pool
+
+        Args:
+
+        Returns:
+            dict with stats response
+        """
+        pool_id = unquote(request.match_info['pool_id'])
+
+        if pool_id is None:
+            return web.json_response(
+                {}, dumps=safe_json_dumps, status=400)
+        pool_id = int(pool_id)
+        data, status = await BaseHandler.resp_query_worker_stats(pool_id)
+        return web.json_response(data, dumps=safe_json_dumps, status=status)
 
     @classmethod
     @EnodoAuth.auth.required
@@ -93,110 +123,6 @@ class ApiHandlers:
                 data['query'])
         return web.json_response(
             resp, dumps=safe_json_dumps, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    @implement_fields_query
-    async def get_series_configs(cls, request, fields=None):
-        """Get all series config templates
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-
-        Query args:
-            fields (String, comma seperated): list of fields to return
-        """
-        data, status = BaseHandler.resp_get_series_configs(
-            fields=fields)
-        return web.json_response(
-            data=data, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def add_series_config(cls, request):
-        """Add a series config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        try:
-            data = await request.json()
-        except JSONDecodeError as e:
-            resp, status = {'error': 'Invalid JSON'}, 400
-        else:
-            resp, status = await BaseHandler.resp_add_series_config(
-                data)
-        return web.json_response(data=resp, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def remove_series_config(cls, request):
-        """Remove a series config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        rid = urllib.parse.unquote(
-            request.match_info['rid'])
-        data, status = await BaseHandler.resp_remove_series_config(
-            rid)
-        return web.json_response(
-            data=data, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def update_series_config_static(cls, request):
-        """Update a series config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        try:
-            data = await request.json()
-        except JSONDecodeError as e:
-            resp, status = {'error': 'Invalid JSON'}, 400
-        rid = urllib.parse.unquote(
-            request.match_info['rid'])
-        data, status = \
-            await BaseHandler.resp_update_series_config_static(
-                rid, data.get("name"), data.get("description"))
-        return web.json_response(
-            data=data, status=status)
-
-    @classmethod
-    @EnodoAuth.auth.required
-    async def update_series_config(cls, request):
-        """Update a series config template
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        try:
-            data = await request.json()
-        except JSONDecodeError as e:
-            resp, status = {'error': 'Invalid JSON'}, 400
-        rid = urllib.parse.unquote(
-            request.match_info['rid'])
-        data, status = \
-            await BaseHandler.resp_update_series_config(
-                rid, data)
-        return web.json_response(
-            data=data, status=status)
 
     @classmethod
     @EnodoAuth.auth.required
@@ -331,20 +257,6 @@ class ApiHandlers:
 
     @classmethod
     @EnodoAuth.auth.required
-    async def get_event_log(cls, request):
-        """Returns enodo event log
-
-        Args:
-            request (Request): aiohttp request
-
-        Returns:
-            _type_: _description_
-        """
-        log = ""
-        return web.json_response(data={'data': log}, status=200)
-
-    @classmethod
-    @EnodoAuth.auth.required
     @implement_fields_query
     async def get_settings(cls, request, fields=None):
         """Returns current settings dict
@@ -419,7 +331,27 @@ class ApiHandlers:
         resp = await BaseHandler.resp_get_enodo_stats(fields=fields)
         return web.json_response(data=resp, status=200)
 
-    
+    @classmethod
+    @EnodoAuth.auth.required
+    async def get_workers(cls, request):
+        """Get workers in a pool
+
+        Args:
+            request (Request): aiohttp request
+
+        Returns:
+            _type_: _description_
+        """
+        try:
+            pool_id = int(request.match_info['pool_id'])
+        except Exception:
+            return web.json_response(
+                data={"error": "Invalid pool_id"},
+                status=400)
+
+        status, resp = BaseHandler.resp_get_workers(pool_id)
+        return web.json_response(data=resp, status=status)
+
     @classmethod
     @EnodoAuth.auth.required
     async def add_worker(cls, request):
@@ -431,6 +363,53 @@ class ApiHandlers:
         Returns:
             _type_: _description_
         """
+        try:
+            pool_id = int(request.match_info['pool_id'])
+        except Exception:
+            return web.json_response(
+                data={"error": "Invalid pool_id"},
+                status=400)
+
+        try:
+            data = await request.json()
+        except Exception:
+            return web.json_response(
+                data={"error": "Invalid config in payload, invalid JSON"},
+                status=400)
+        if not isinstance(data, dict):
+            return web.json_response(
+                data={"error": "Invalid config in payload"},
+                status=400)
+
         data = await request.json()
-        status, resp = await BaseHandler.resp_add_worker(data)
+        status, resp = await BaseHandler.resp_add_worker(pool_id, data)
+        return web.json_response(data=resp, status=status)
+
+    @classmethod
+    @EnodoAuth.auth.required
+    async def delete_worker(cls, request):
+        """Delete worker
+
+        Args:
+            request (Request): aiohttp request
+
+        Returns:
+            _type_: _description_
+        """
+        try:
+            pool_id = int(request.match_info['pool_id'])
+        except Exception:
+            return web.json_response(
+                data={"error": "Invalid pool_id"},
+                status=400)
+
+        try:
+            job_type_id = int(request.match_info['job_type_id'])
+        except Exception:
+            return web.json_response(
+                data={"error": "Invalid job_type_id"},
+                status=400)
+
+        status, resp = await BaseHandler.resp_delete_worker(
+            pool_id, job_type_id)
         return web.json_response(data=resp, status=status)